divide driver.ts & add script typedoc
